+++ libs/cdk/src/lib/components/forms/components/select/components/combobox/directives/public-api.ts
@@ -1,1 +1,3 @@
+export * from './combobox-option-template/public-api';
+export * from './combobox-selected-option-template/public-api';
 export * from './combobox/public-api';