+++ libs/cdk/src/lib/components/forms/components/select/components/combobox/stories/combobox.component.stories.ts
@@ -80,6 +80,7 @@ export default {
     placeholder: 'Select an option',
     bindLabel: 'name',
     bindValue: 'id',
+    _customOptionTemplate: false,
   },
 } as Meta<StorybookComboboxComponent>;