+++ libs/cdk/src/lib/components/forms/components/select/components/combobox/stories/components/combobox-storybook.component.ts
@@ -1,5 +1,12 @@
-import { JsonPipe } from '@angular/common';
-import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
+import { JsonPipe, NgIf } from '@angular/common';
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  ViewEncapsulation,
+  inject,
+  signal,
+} from '@angular/core';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { ComboboxImports } from '../../../../..';
 
@@ -20,7 +27,16 @@ import { ComboboxImports } from '../../../../..';
         [placeholder]="placeholder"
         [allowCustomValues]="allowCustomValues"
         [filterInternal]="filterInternal"
-      />
+      >
+        <ng-container *ngIf="customOptionTemplate()">
+          <ng-template etComboboxOptionTemplate let-option="option">
+            {{ option.name || option }} (Custom tpl)
+          </ng-template>
+          <ng-template etComboboxSelectedOptionTemplate let-option="option">
+            {{ option.name || option }} (Custom tpl)
+          </ng-template>
+        </ng-container>
+      </et-combobox>
     </et-select-field>
 
     <pre> {{ fg.value | json }} </pre>
@@ -28,9 +44,23 @@ import { ComboboxImports } from '../../../../..';
   standalone: true,
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
-  imports: [ComboboxImports, ReactiveFormsModule, JsonPipe],
+  styles: [
+    `
+      et-sb-combobox {
+        display: block;
+        padding-top: 150px;
+      }
+
+      pre {
+        padding-bottom: 100px;
+      }
+    `,
+  ],
+  imports: [ComboboxImports, ReactiveFormsModule, JsonPipe, NgIf],
 })
 export class StorybookComboboxComponent {
+  private readonly _cdr = inject(ChangeDetectorRef);
+
   fg = new FormControl({ value: null, disabled: false });
 
   options: unknown[] = [];
@@ -56,4 +86,14 @@ export class StorybookComboboxComponent {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     this.fg.setValue(value as any);
   }
+
+  set _customOptionTemplate(value: boolean) {
+    this.customOptionTemplate.set(value);
+
+    setTimeout(() => {
+      this._cdr.markForCheck();
+    }, 1);
+  }
+
+  customOptionTemplate = signal(false);
 }