+++ libs/cdk/src/lib/components/overlay/components/tooltip/stories/tooltip.stories.ts
@@ -1,5 +1,5 @@
-import { placements } from '@popperjs/core';
 import { Meta, Story } from '@storybook/angular';
+import { FLOATING_UI_PLACEMENTS } from '../../../../../utils';
 import { TooltipStorybookComponent } from './components';
 import CustomMDXDocumentation from './tooltip.docs.mdx';
 
@@ -16,11 +16,11 @@ export default {
       control: {
         type: 'select',
       },
-      options: placements,
+      options: FLOATING_UI_PLACEMENTS,
     },
   },
   args: {
-    placement: 'auto',
+    placement: 'bottom',
   },
 } as Meta<TooltipStorybookComponent>;