+++ libs/cdk/src/lib/public-api.ts
@@ -1,3 +1,4 @@
 export * from './components/public-api';
 export * from './services/public-api';
 export * from './types/public-api';
+export * from './utils/public-api';