+++ libs/query/src/lib/devtools/public-api.ts
@@ -1,0 +1,3 @@
+export * from './components/public-api';
+export * from './pipes/public-api';
+export * from './utils/public-api';